+++ ZeroMev/MevEFC/Block.cs
@@ -5,7 +5,7 @@ namespace ZeroMev.MevEFC
 {
     public partial class Block
     {
-        public decimal BlockNumber { get; set; }
+        public long BlockNumber { get; set; }
         public DateTime BlockTimestamp { get; set; }
     }
 }
