+++ ZeroMev/MevEFC/NftTrade.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
 
 namespace ZeroMev.MevEFC
 {
@@ -8,16 +9,16 @@ namespace ZeroMev.MevEFC
         public DateTime? CreatedAt { get; set; }
         public string AbiName { get; set; } = null!;
         public string TransactionHash { get; set; } = null!;
-        public decimal TransactionPosition { get; set; }
-        public decimal BlockNumber { get; set; }
+        public int TransactionPosition { get; set; }
+        public long BlockNumber { get; set; }
         public string TraceAddress { get; set; } = null!;
         public string Protocol { get; set; } = null!;
         public string? Error { get; set; }
         public string SellerAddress { get; set; } = null!;
         public string BuyerAddress { get; set; } = null!;
         public string PaymentTokenAddress { get; set; } = null!;
-        public decimal PaymentAmount { get; set; }
+        public BigInteger PaymentAmount { get; set; }
         public string CollectionAddress { get; set; } = null!;
-        public decimal TokenId { get; set; }
+        public BigInteger TokenId { get; set; }
     }
 }
