+++ ZeroMev/MevEFC/ZmBlock.cs
@@ -5,7 +5,7 @@ namespace ZeroMev.MevEFC
 {
     public partial class ZmBlock
     {
-        public decimal BlockNumber { get; set; }
+        public long BlockNumber { get; set; }
         public int TransactionCount { get; set; }
         public DateTime? BlockTime { get; set; }
         public byte[]? TxData { get; set; }
