+++ ZeroMev/MevEFC/ZmLatestBlockUpdate.cs
@@ -5,7 +5,7 @@ namespace ZeroMev.MevEFC
 {
     public partial class ZmLatestBlockUpdate
     {
-        public decimal BlockNumber { get; set; }
+        public long BlockNumber { get; set; }
         public DateTime? UpdatedAt { get; set; }
     }
 }
